implement runMode in all modules

+++ core/src/main/scala/cpslab/deploy/CommonUtils.scala
@@ -60,8 +60,8 @@ object CommonUtils {
                                            maxRangeNum: Int): List[LoadData] = {
     // TODO: limit the minimum range
     val loadDataRequests = new ListBuffer[LoadData]
-    val startRowKeyInt = Bytes.toLong(startRowKey)
-    val endRowKeyInt = Bytes.toLong(endRowKey)
+    val startRowKeyInt = Bytes.toInt(startRowKey)
+    val endRowKeyInt = Bytes.toInt(endRowKey)
     var newStartKeyInt = startRowKeyInt
     val maxLength = (endRowKeyInt - startRowKeyInt + 1) / maxRangeNum
     while (newStartKeyInt <= endRowKeyInt) {
